Another rewrite of how the whole thing works
The shop url doesn't always match the 'myshopify.com' shop name, so
previously was loading in data for both the store url and the shop name
manually, on a per store basis.
Update to pull in the store name from the loaded page so this should now
work with any Shopify store.